Formatter: Add `--format` option, with JSON output
This adds the ability to output all reports as a single json string.
